侵入式修改cppzmq的zmq_addon.hpp以避免编译器警告
Desc: 侵入式修改cppzmq，根据上下文（重点可关注228行）可以确定该处必须为uint32_t类型，该处在x64下会导致编译器报warning，侵入式修改避免编译器警告。
Type: Feature
Link: NA

+++ 3rd/cppzmq/zmq_addon.hpp
@@ -236,7 +236,7 @@ message_t encode(const Range &parts)
     message_t encoded(mmsg_size);
     unsigned char *buf = encoded.data<unsigned char>();
     for (const auto &part : parts) {
-        const uint32_t part_size = part.size();
+        const uint32_t part_size = static_cast<uint32_t>(part.size());
         const unsigned char *part_data =
           static_cast<const unsigned char *>(part.data());
